integration tests: Extract MockSocket to its own file
