+++ daemon/integration-tests/src/lib.rs
@@ -4,3 +4,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 pub mod actors;
+pub mod socket;
